Update ru_ru.json
It was not necessary to add "%s" to the text as it does not translate to Russian, additional localization strings could have been added, and integration with other mods could have been added without any issues. Furthermore, it does not look good on your part + the decimal point is not displayed properly (highlighted in red).